move the addition of tracking url from various places to only at the end to html_string
