+++ templates/streams/newsletter/content_block.html
@@ -81,7 +81,7 @@
                 <td align="left" style="font-size:0px;padding:10px 25px;padding-bottom:0;word-break:break-word;">
                   <div
                     style="font-family:futura-pt, Arial, sans-serif;font-size:21px;font-style:normal;font-weight:400;line-height:21px;text-align:left;color:#000000;">
-                    <a href="{{ url|slice:':-1' }}?utm_source=cigi_newsletter&utm_medium=email&utm_campaign={{page.title|slugify}}" style="color: #000000; text-decoration: none;">
+                    <a href="{{ url|slice:':-1' }}" style="color: #000000; text-decoration: none;">
                       {{ title }}
                     </a>
                   </div>
@@ -115,7 +115,7 @@
                       <td align="center" bgcolor="transparent" role="presentation"
                         style="border:1px solid #e71b3e;border-radius:2px;cursor:auto;font-style:normal;mso-padding-alt:8px 16px;background:transparent;"
                         valign="middle">
-                        <a href="{{ url|slice:':-1' }}?utm_source=cigi_newsletter&utm_medium=email&utm_campaign={{page.title|slugify}}"
+                        <a href="{{ url|slice:':-1' }}"
                           style="display:inline-block;background:transparent;color:#e71b3e;font-family:futura-pt, Arial, sans-serif;font-size:10px;font-style:normal;font-weight:400;line-height:1;margin:0;text-decoration:none;text-transform:none;padding:8px 12px;mso-padding-alt:0px;border-radius:2px;"
                           target="_blank">
                           {% if cta_image_link %}<img src="{{ cta_image_link }}" width="13" height="13"
